Updates from review, make newList plain list, update contact form list
