+++ src/contacts/view/ContactView.ts
@@ -52,7 +52,7 @@ import { exportAsVCard, importAsVCard } from "./ImportAsVCard.js"
 import { MobileHeader } from "../../gui/MobileHeader.js"
 import { LazySearchBar } from "../../misc/LazySearchBar.js"
 import { MultiselectMobileHeader } from "../../gui/MultiselectMobileHeader.js"
-import { MultiselectMode } from "../../gui/base/NewList.js"
+import { MultiselectMode } from "../../gui/base/List.js"
 import { EnterMultiselectIconButton } from "../../gui/EnterMultiselectIconButton.js"
 import { selectionAttrsForList } from "../../misc/ListModel.js"
 import { ContactViewModel } from "./ContactViewModel.js"
