+++ src/mail/view/MailView.ts
@@ -56,7 +56,7 @@ import { LazySearchBar } from "../../misc/LazySearchBar.js"
 import { MultiselectMobileHeader } from "../../gui/MultiselectMobileHeader.js"
 import { MailViewModel } from "./MailViewModel.js"
 import { selectionAttrsForList } from "../../misc/ListModel.js"
-import { ListLoadingState, MultiselectMode } from "../../gui/base/NewList.js"
+import { ListLoadingState, MultiselectMode } from "../../gui/base/List.js"
 import { EnterMultiselectIconButton } from "../../gui/EnterMultiselectIconButton.js"
 import { MailFilterButton } from "./MailFilterButton.js"
 import { listSelectionKeyboardShortcuts } from "../../gui/base/ListUtils.js"
