+++ src/mail/view/MailViewModel.ts
@@ -4,11 +4,10 @@ import { EntityClient } from "../../api/common/EntityClient.js"
 import { Mail, MailFolder, MailTypeRef } from "../../api/entities/tutanota/TypeRefs.js"
 import { firstBiggerThanSecond, GENERATED_MAX_ID, getElementId, isSameId, sortCompareByReverseId } from "../../api/common/utils/EntityUtils.js"
 import { assertNotNull, count, debounce, lastThrow, lazyMemoized, memoized, ofClass, promiseFilter } from "@tutao/tutanota-utils"
-import { ListState } from "../../gui/base/NewList.js"
+import { ListState } from "../../gui/base/List.js"
 import { ConversationPrefProvider, ConversationViewModel, ConversationViewModelFactory } from "./ConversationViewModel.js"
 import { CreateMailViewerOptions } from "./MailViewer.js"
 import { EntityUpdateData, EventController, isUpdateForTypeRef } from "../../api/main/EventController.js"
-import { ListFetchResult } from "../../gui/base/List.js"
 import { isOfflineError } from "../../api/common/utils/ErrorCheckUtils.js"
 import { MailFolderType } from "../../api/common/TutanotaConstants.js"
 import { assertSystemFolderOfType, isOfTypeOrSubfolderOf, isSpamOrTrashFolder, isSubfolderOfType } from "../../api/common/mail/CommonMailUtils.js"
@@ -23,6 +22,7 @@ import { getMailFilterForType, MailFilterType } from "../model/MailUtils.js"
 import Stream from "mithril/stream"
 import { InboxRuleHandler } from "../model/InboxRuleHandler.js"
 import { Router } from "../../gui/ScopedRouter.js"
+import { ListFetchResult } from "../../gui/base/ListUtils.js"
 
 export interface MailOpenedListener {
 	onEmailOpened(mail: Mail): unknown
