+++ src/search/SearchBar.ts
@@ -9,17 +9,16 @@ import { ContactTypeRef, MailTypeRef } from "../api/entities/tutanota/TypeRefs.j
 import type { Shortcut } from "../misc/KeyManager"
 import { keyManager } from "../misc/KeyManager"
 import { NotAuthorizedError, NotFoundError } from "../api/common/error/RestError"
-import { getRestriction, getSearchUrl, isAdministratedGroup, setSearchUrl } from "./model/SearchUtils"
+import { getRestriction } from "./model/SearchUtils"
 import { locator } from "../api/main/MainLocator"
 import { Dialog } from "../gui/base/Dialog"
-import type { GroupInfo, WhitelabelChild } from "../api/entities/sys/TypeRefs.js"
+import type { WhitelabelChild } from "../api/entities/sys/TypeRefs.js"
 import { GroupInfoTypeRef, WhitelabelChildTypeRef } from "../api/entities/sys/TypeRefs.js"
 import { FULL_INDEXED_TIMESTAMP, Keys } from "../api/common/TutanotaConstants"
 import { assertMainOrNode, isApp } from "../api/common/Env"
 import { styles } from "../gui/styles"
 import { client } from "../misc/ClientDetector"
 import { debounce, downcast, flat, groupBy, isSameTypeRef, memoized, mod, ofClass, promiseMap, TypeRef } from "@tutao/tutanota-utils"
-import { PageSize } from "../gui/base/List"
 import { BrowserType } from "../misc/ClientConstants"
 import { hasMoreResults } from "./model/SearchModel"
 import { SearchBarOverlay } from "./SearchBarOverlay"
@@ -31,6 +30,7 @@ import { compareContacts } from "../contacts/view/ContactGuiUtils"
 import { LayerType } from "../RootView"
 import { BaseSearchBar, BaseSearchBarAttrs } from "../gui/base/BaseSearchBar.js"
 import { SearchRouter } from "./view/SearchRouter.js"
+import { PageSize } from "../gui/base/ListUtils.js"
 
 assertMainOrNode()
 export type ShowMoreAction = {
