+++ test/tests/Suite.ts
@@ -48,7 +48,6 @@ import "./misc/FormatterTest.js"
 import "./api/worker/UrlifierTest.js"
 import "./misc/PasswordUtilsTest.js"
 import "./gui/animation/AnimationsTest.js"
-import "./gui/base/ListTest.js"
 import "./gui/ThemeControllerTest.js"
 import "./api/main/EntropyCollectorTest.js"
 import "./misc/HtmlSanitizerTest.js"
